+++ lib/staged_event.rb
@@ -2,10 +2,10 @@
 
 require "active_record"
 require "directive"
+require_relative "staged_event/backoff_timer"
 require_relative "staged_event/configuration"
 require_relative "staged_event/message_envelope_pb"
 require_relative "staged_event/model"
-require_relative "staged_event/publisher/backoff_timer"
 require_relative "staged_event/publisher/base"
 require_relative "staged_event/publisher/google_pub_sub"
 require_relative "staged_event/publisher_process"
